Refactoring due to the comments
