+++ src/haystack_integrations/components/rankers/voyage/__init__.py
@@ -1,0 +1,3 @@
+from haystack_integrations.components.rankers.voyage.ranker import VoyageRanker
+
+__all__ = ["VoyageRanker"]
