+++ src/haystack_integrations/components/rankers/voyage/ranker.py
