--- src/haystack_integrations/components/rerankers/voyage_rerankers/__init__.py
@@ -1,3 +1,0 @@
-from haystack_integrations.components.rerankers.voyage_rerankers.voyage_text_reranker import VoyageRanker
-
-__all__ = ["VoyageRanker"]
